Extract shared domain file contracts
