+++ src/CaptureTool.Application.Abstractions/Features/AudioEdit/OpenAudioEditPage/OpenAudioEditPageRequest.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Application.Abstractions.Features.AudioEdit.OpenAudioEditPage;
