+++ src/CaptureTool.Application.Abstractions/Features/ImageEdit/OpenImageEditPage/OpenImageEditPageRequest.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Application.Abstractions.Features.ImageEdit.OpenImageEditPage;
