+++ src/CaptureTool.Application.Abstractions/Features/VideoEdit/OpenVideoEditPage/OpenVideoEditPageRequest.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Application.Abstractions.Features.VideoEdit.OpenVideoEditPage;
