+++ src/CaptureTool.Application.Abstractions/Storage/IFilePickerService.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 using System.Drawing;
 
 namespace CaptureTool.Application.Abstractions.Storage;
