+++ src/CaptureTool.Application/Features/AppMenu/OpenFile/OpenFileUseCase.cs
@@ -5,7 +5,7 @@ using CaptureTool.Application.Abstractions.Files;
 using CaptureTool.Application.Abstractions.Navigation;
 using CaptureTool.Application.Abstractions.Storage;
 using CaptureTool.Domain.Capture;
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 using CaptureTool.Application.Abstractions.UseCases;
 using CaptureTool.Application.Features.AudioCapture;
