+++ src/CaptureTool.Domain.Capture/VideoFile.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Domain.Capture;
