+++ src/CaptureTool.Domain/CaptureTool.Domain.csproj
@@ -1,0 +1,1 @@
+<Project Sdk="Microsoft.NET.Sdk" />
