+++ src/CaptureTool.Domain/Files/IAudioFile.cs
@@ -1,4 +1,4 @@
-namespace CaptureTool.Domain.Capture.Files;
+namespace CaptureTool.Domain.Files;
 
 public partial interface IAudioFile : IFile
 {
