+++ src/CaptureTool.Infrastructure.Capture.Windows/WindowsAudioRecorder.cs
@@ -1,6 +1,6 @@
 using CaptureKit.Abstractions;
 using CaptureTool.Domain.Capture;
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Infrastructure.Capture.Windows;
