+++ src/CaptureTool.Infrastructure.Windows/Storage/WindowsFile.cs
@@ -1,4 +1,4 @@
-using CaptureTool.Domain.Capture.Files;
+using CaptureTool.Domain.Files;
 
 namespace CaptureTool.Infrastructure.Windows.Storage;
